revert to v1.47.3 for golangci-lint + some refactoring
