+++ acceptance_test.go
@@ -20,6 +20,7 @@ import (
 
 	"cloud.google.com/go/storage"
 	"github.com/conduitio/conduit-connector-google-cloudstorage/config"
+	"github.com/conduitio/conduit-connector-google-cloudstorage/source/utils"
 	sdk "github.com/conduitio/conduit-connector-sdk"
 	"github.com/google/uuid"
 	"go.uber.org/goleak"
@@ -48,12 +49,12 @@ func (d GCSAcceptanceTestDriver) WriteToSource(t *testing.T, records []sdk.Recor
 }
 
 func TestAcceptance(t *testing.T) {
-	sourceConfig, err := parseIntegrationConfig()
+	sourceConfig, err := utils.ParseIntegrationConfig()
 	if err != nil {
 		t.Skip(err)
 	}
 
-	gcsClient, err := newGCSClient(sourceConfig)
+	gcsClient, err := utils.NewGCSClient(sourceConfig)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -72,12 +73,12 @@ func TestAcceptance(t *testing.T) {
 				DestinationConfig: nil,
 				BeforeTest: func(t *testing.T) {
 					sourceConfig[config.ConfigKeyGCSBucket] = "acceptance-test-bucket-" + uuid.NewString()
-					if err := createTestGCSBucket(gcsClient, sourceConfig["projectID"], sourceConfig[config.ConfigKeyGCSBucket]); err != nil {
+					if err := utils.CreateTestGCSBucket(gcsClient, sourceConfig["projectID"], sourceConfig[config.ConfigKeyGCSBucket]); err != nil {
 						t.Fatalf("could not create test gcs bucket: %v", err)
 					}
 				},
 				AfterTest: func(t *testing.T) {
-					clearAndDeleteTestGCSBucket(t, gcsClient, sourceConfig[config.ConfigKeyGCSBucket])
+					utils.ClearAndDeleteTestGCSBucket(t, gcsClient, sourceConfig[config.ConfigKeyGCSBucket])
 				},
 				// Apart from the IgnoreCurrent, runtime_pollWait is also ignorned because here the GCS/storage client(Created Above) make a gRPC connection which is consistent and opens until it is closed.
 				GoleakOptions: []goleak.Option{goleak.IgnoreCurrent(), goleak.IgnoreTopFunction("internal/poll.runtime_pollWait")},
